No address fields in config yet
