+++ api/src/main/java/org/openmrs/module/pihcore/config/registration/RegistrationConfigDescriptor.java
@@ -11,9 +11,6 @@ import java.util.Map;
 public class RegistrationConfigDescriptor {
 
     @JsonProperty
-    private AddressConfigDescriptor addressConfig;
-
-    @JsonProperty
     private boolean allowUnknownPatients = false;
 
     @JsonProperty
@@ -39,14 +36,6 @@ public class RegistrationConfigDescriptor {
 
     @JsonProperty
     private Integer maxPatientMatchResults;
-
-    public AddressConfigDescriptor getAddressConfig() {
-        return addressConfig;
-    }
-
-    public void setAddressConfig(AddressConfigDescriptor addressConfig) {
-        this.addressConfig = addressConfig;
-    }
 
     public boolean isAllowUnknownPatients() {
         return allowUnknownPatients;
